requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-csp==4.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
filelock==3.16.1
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Markdown==3.7
MouseInfo==0.1.3
mysql-connector-python==9.1.0
numpy==2.1.3
opencv-contrib-python==4.10.0.84
packaging==24.2
pillow==11.1.0
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
requests==2.32.3
setuptools==75.8.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
uvicorn==0.34.0
virtualenv==20.29.1
whitenoise==6.9.0
